pyproject.toml
[project]
name = "script-youtube-s3"
version = "0.1.0"
description = "Для работы с youtube видео через s3"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.39.3",
    "python-decouple>=3.8",
    "yt-dlp>=2025.6.30",
]
[tool.hatch.build.targets.wheel]
packages = ["main.py"]

[project.scripts]
download = "main:download"
upload = "main:upload"
youtube-s3 = "main:youtube_to_s3"
list-bucket = "main:list_bucket"
check-config = "main:check_config"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"